Add a test that covers out-of-order `require_auth` calls. (#703)
'Out-of-order' here means that the parent contract's `require_auth` is called after child's `require_auth`. Initially I thought that might be problematic, but it seems to work consistently across both recording and enforcing modes: the calls trees are created as soon as the first `require_auth` for and address happens and they are no longer usable after the respective call is popped out of call stack. So for this case we simply would end up with two trees. This should normally not be happen, but there is really nothing wrong with this case.