Introduce temporary storage in the host.
Temporary storage is basically a contract-owned map that isn't persisted in the ledger, but may be used to share values between multiple contract calls within the same top-level invocation. This is useful to implement safe allowance-style patterns without the risk of the allowance to stay in the ledger indefinitely (e.g. to transfer token to a dynamic, arbitrary number of recipients).