Attach the auth in recording mode to any valid tracker.
Currently only the tracker that matches with exact same object is used and otherwise a new tracker is created. This results in recording auth producing unnecessarily disjoint call trees.

This is kind of best effort and might lead to a bit weird edge-case scenarios in semantic sense (e.g. a tracker for an exact address object might be used by another instance of that address), but the end results should still fulfil the enforcing mode invariants and shouldn't compromise security.

Also improved the test coverage to cover the case where different objects need to be mapped to the same auth tree, as well as additional coverage for duplicate address auth.