Env-common cleanups and dead code removal (#1077)
A bunch more improvements from code review of common: convert, env,
error, lib, meta:

- Remove various sources of potential panic: calls to panic!,
unreachable!, assert! or the like.
- Weaken the recently-added From<Error> bound on EnvBase::Error, move it
to a method provided by the Env, to remove the ambient impl of
From<Error> for Infallible that was in common. Only the Guest needs to
provide such a path.
- Remove some other dead methods in EnvBase.
- Remove the dead unimplemented_env and the non-wasm Guest definition.
- Improve some comments and localize some cfgs.

Also includes a cherry pick of f03b54212de5355680f49be023fb950371468b80
that bumps the version to rc2 (this seems not to have landed on main,
just the tag?)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>