Remove bad LEs (#1114)
This comes out of some review on compare.rs: we were supporting a lot of
LE types in the code that are not ever supposed to occur. It's better to
treat the presence of those cases as internal errors than have dead code
"supporting" them possibly incorrectly.

Also eliminates comparison of LEs altogether. They were only being
compared because of older constraints on MeteredOrdMap wanting its value
type to be comparable. That's not necessary for LEs (in the Storage
map), only Vals (in HostMaps) in order to have Map be ordered itself
(because Val must be ordered). Turns out we'd already relaxed the
requirement on MeteredOrdMap's value type to not have to be comparable.
So all the LE-ordering code was dead anyway, which is good because they
were all being treated as calls to Ord on fixed size types, but they
weren't all fixed size types.

Also fixed a missing case in metered_clone, where we weren't charging
for substructure of AccountEntries.

I'm not sure this is all that needs to be done for these files. I'm just
out of time for tonight. Will continue reviewing on monday but figured I
should get this started now.