Review fixes for declared_size and metered_clone (#1120)
Fair amount of churn here:

- Removed several dead impls, cleaned up comments, added missing
testcases, reordered stuff for consistency
- Reclassified and removed incorrectly-classified shallow types. This is
fairly serious: we weren't counting the substructure cloning cost of
much of the auth system as well as the wasm code of a contractcode
entry, or the signers of an account entry.

One interesting part of this is that, as far as I can tell, everything
in the events system doesn't use `metered_clone` anymore at all. I think
this is good because it previously _errored_ when trying to
`metered_clone` a diagnostic event, which is "just as observable" as
`metered_cloning`-something-that-should-be-free would have been. Luckily
it looks like we only "accumulate events", don't clone them, and then
externalize them when done, charging the cost of emitting them outside
and ignoring the cost of externalizing. I think! @jayz22 and @sisuresh I
would appreciate a second pair of eyes on this understanding.