Cackle gate (#1124)
This adds a transitive-API-usage-checker called
[`cackle`](https://davidlattimore.github.io/making-supply-chain-attacks-harder.html)
to the soroban build. It is currently (with a few exceptions) wired to
prohibit most uses of the filesystem, network, time, the environment,
threading, random numbers or hashtables. It's configured by cackle.toml
and there's a new Makefile and github action target for it.

I also took the opportunity to eliminate all uses of HashMap and
thread_rng (see #810).

Thanks to @SirTyson for suggesting this tool! It was a little tricky to
get working right but once working it seems good.