Overhaul "free budget" and migrate non-metered code (#1123)
### What

Resolves https://github.com/stellar/rs-soroban-env/issues/1061

- Overhaul the "free budget":
- Add two fields `internal_limit` and `internal_total_count` to budget
dimension for keeping track of internal work being done, which don't get
metered for fee, but need to be limited for DOS prevention.
- Remove the `with_free_budget` interface and replace it with
`with_internal_mode` (and a `with_debug_budget` on the host side that
wraps around it and asserts host is in `DEBUG` mode), taking two
closures and ensuring 1. the `internal_limit` is not exceeded 2. any
error occurring within is suppressed. The error suppression part is to
prevent error-based logic existing outside the closure that accidentally
causes divergences between debug and no-debug workflows.
     - Adds a new hostile contract case for testing.

- A refactoring to split the `budget.rs` into multiple smaller files.
The file was getting big (1200+ lines) and hard to navigate.
- Migrate all existing non-metered code to metered version, this
includes diagnostic logging, diagnostic event, auth preflight code.

### Why

[TODO: Why this change is being made. Include any context required to
understand the why.]

### Known limitations

[TODO or N/A]

---------

Co-authored-by: Graydon Hoare <graydon@pobox.com>