Add conceptual overview comment to auth module (#1128)
To help prepare myself for reviewing the auth module, I decided it'd be
best to do a careful read of the CAP and XDR and write down my
understanding from a very high level without reading much of the code
beyond data structures. This is the result. I'd love to know if (a) it's
approximately correct and (b) this is a good place to leave the comment,
or if I ought to move it to docs somewhere else (eg. try to incorporate
it into
https://soroban.stellar.org/docs/fundamentals-and-concepts/authorization).