Add a test that ensures that SAC reentry is not possible. (#1179)
### What

Add a test that ensures that SAC reentry is not possible.

The test uses a custom account that does a token transfer in its
`__check_auth`, which causes reentry in case if it's called during the
operation on that token.

### Why

Improving test coverage
(https://github.com/stellar/rs-soroban-env/issues/1166)

### Known limitations

N/A