Remove vnext test wasms (#1181)
### What

Remove vnext test wasms to allow core to test multiple protocol
versions. This only works at the moment because V20 test wasms are
compatible with V21, but this may not be the case in the future. At some
point we may need test wasms organized by protocol version (not just
curr vs vnext).