Initial cut at bug 872, observe side effects of tests (#1142)
This is still a draft but I'd say it's about 75% done -- the main tricky
(or _hacky_) parts are all working about as well as I can make them.

The goal here is to provide something _like_ what we do for the txmeta
recording in core: leverage the unit test suite to record traces of
observable side effects into a stable file checked in to version control
such that we _notice_ when we unintentionally change observable
behaviour of anything the tests exercise.

I just did some preliminary observations to get it working (budget
values and prng state). Will add more.

This doesn't work _quite_ as seamlessly as the core version: the Rust
unit test harness is basically impossible to hook a pre-run / post-run
interceptor into directly, so I wound up providing a wrapper around the
Host that records its state when destroyed, and then augmenting it with
a (macro-captured) name-of-the-function it's created in. Since most of
our tests have exactly one host, this works (and it is only a little
work to generalize to tests with multiple hosts).

Unfortunately there's no shutdown system in Rust at all, nor for the
overall test harness, and the community consensus seems to be that
hooking into libc's `atexit` is the best approach. So I did that too.
Gross.

There are a few things left to do here which I'll have to pick up
tomorrow, but it seems to _mostly_ work:

- [ ] support both curr and next files
- [ ] go through the rest of the testsuite modifying the Host
construction sites
- [x] observe a lot more state about the Host
- [ ] maybe warn when there are unused entries in the observations file?
maybe not since running single tests is common
- [x] connect to the host lifecycle hooks system to do some _cumulative_
observations (eg. as VMs are set up and torn down)