Don't block merges on rust-analyzer compat check (#1204)
### What
Don't block merges on rust-analyzer compat check.

### Why
Lately it's taking a long time to run, longer than other jobs in the
workflows.

It's been more than 6 months I think since we broke this compat check.
We might not need it anymore. We just don't seem to be stumbling upon
limitations in rust-analyzer, we might mean the things we were stumbling
on in the past have been fixed. We could remove it. Simply stopping it
from blocking merges is a step towards doing that.

Related:
- https://github.com/stellar/rs-soroban-sdk/pull/1151