Fix #1175 error code spoofing (#1209)
I _think_ this is actually all that we need (or should) do for #1175.
It's a little less than I was expecting to change: in particular I kept
the adjustment in `with_frame`, and just tightened its condition (and
commented why).

There are more call paths that go through `with_frame` than I expected
and adjusting all of them is a bit more repetitive and introduces a bit
more risk of drift between them than I think warrants the risk of
accidentally over-adjusting an error unintentionally here (eg. if
someone uses with_frame to pass an error we _don't_ want adjusted this
way for some reason).

Nothing is ideal in this story, but I think this is a conservative fix
that at least avoids the problem we were concerned with. @dmkozh wdyt? I
will tack on a test that exercises this if you concur.