Do not error if no authorizations (#1214)
### What
When asking the Host for a list of authenticated authorizations do not
generate an error if there are none.

### Why
At the moment when asking the host for information about any auths that
have been seen it generates and records an error if no auth was setup.
This is inconvenient because this function is primarily called by tests,
but the generation and recording of the error goes into the diagnostic
log and muddies the logs for a developer. It makes it look like the
contract had some problem where-as it was simply a test utility called
at a time when it had no data.

There's little benefit in returning an error when no auth exists.
Instead we can simply return an empty list. No auth was setup, no auth
occurred, therefore an empty list showing that no auth occurred can be
returned.

I encountered this as a problem when trying to store auth snapshots in
the Rust SDK. At the end of a test I ask the host for auth info, but
that causes a diagnostic error to be written into the events for every
test that does nothing with auth.