Observe more things (#1224)
This adds several new observation types to the test observer. The most
important part is that it now observes the state of the auth manager,
which is reports with two separate values: a (size,hash) pair for the
current auth stack, and a (size,hash) pair for the sum of all the
trackers. I _think_ this is probably the most useful way to eyeball
changes, but @dmkozh should confirm (and maybe give a look at the
recorded observation traces here to see if they make sense to you)