Add Host::has_frame (#1244)
### What
Add a method to the host for determining if the host currently has a
frame on the stack.

### Why
In several places in the SDK it would be ideal if we could error if a
function is used inside or outside of a contract function, because doing
so would be invalid.

In some code being added or auth tracking in the SDK I need to reliable
know if we are inside a contract function or not.

I tried to do this using top invocation hooks, but the way native
contract functions get called in the SDK doesn't pass through anything
from the Env that is available outside of the contract. We can sort of
know if we're in a contract, but it isn't guaranteed to be correct in
edge cases. We might change that some day, but at the moment that
limitation exists.

Exposing whether the host currently has a frame is a simple way the host
can indicate to SDK test utilities if it is currently in a contract
function or not.

Close https://github.com/stellar/rs-soroban-sdk/issues/1084