Fuzz fixes (#1249)
This contains the results of a few days of both reproducing and fixing
fuzz bugs and also hacking on the fuzzer:

1. Fixes #1203 which was a somewhat subtle budget-error-escalation bug
the fuzzer found.
2. Fixes a couple other instances of that pattern.
3. Fixes a fuzzer-found instance of a frame invariant corruption.
4. Moves the machinery the fuzzer needs from tests/utils (which was
guarded by `cfg(test)` so inaccessible to `testutils` and guards it with
`cfg(feature="testutils")`
5. Removes the duplicate code in the fuzzer lib and instead uses host
testutils
6. Extends testutils to record the storage map and then switch to
enforcing, rather than manually building the enforcing map
7. Removes the other non-expr fuzz case, it's redundant.
8. Adds some scaffolding for fuzz debugging.
9. Adds support for passing arguments to the fuzzer to enable access to
storage and cross-contract calls.
10. Fix the thing where we rebuild every time due to build.rs not being
conservative enough