Noop auth fix (#1278)
### What

The SAC uses the `AuthorizedToMaintainLiabilitiesFlag` to deauthorize a
trustline. If a trustline is completely deauthorized (`AuthorizedFlag`
and `AuthorizedToMaintainLiabilitiesFlag` are both not set) and you call
the deauthorize function from the SAC on it, the SAC will set
`AuthorizedToMaintainLiabilitiesFlag`. This PR makes that scenario (as
well as authorizing an authorized trustline) a noop for the auth flag
state.