Add time tracker (#1279)
### What

Add a time tracker to the budget. 

### Why

Such that we can track the time spent in individual cost type, notably
`VmInstantiation`.
Combined with the cpu cost per type. We can come up with metrics in the
downstream such as "metered cpu per nsecs excluding VM".

### Known limitations

Most of types are not being tracked, and will show as 0. 
We are primarily interested in the `VmInstantiation` which is being
tracked.

---------

Co-authored-by: Graydon Hoare <graydon@pobox.com>