Improve cost model fitting for better lower-range accuracy (#1293)
### What

This PR was originated from the fact that previously the `VarDeser`
cost's constant term was unrealistically high.

It improves the numerical approach towards model generation by
constraining it to the smallest input in the range.
Thus it avoids fitting a model that tries to produce a marginally better
fitness at the expense of lower-range accuracy.

It also removes the dependency of the `HostCostMeasurement`'s input size
on the linear parameter's scale factor
(`COST_MODEL_LIN_TERM_SCALE_BITS`), which previously put a numerical
constrain on the smallest input size and minimal gaps.

[raw calibration
outputs](https://github.com/stellar/rs-soroban-env/files/13621533/output_x86_dec_8.txt)

### Why

[TODO: Why this change is being made. Include any context required to
understand the why.]

### Known limitations

[TODO or N/A]