Add some basic test coverage for `e2e_invoke`. (#1328)
### What

This covers every host fn and some interesting failure scenarios, like
missing footprints/auth.

This also includes a non-observable change to e2e_invoke error reporting
to make tests more consistent: the intention was to return
`InvokeHostFunctionResult` containing an error, but we almost always
would return just a `HostError` directly. Both are processed in the same
way by the Core, thus this is not observable.

### Why

Increasing test coverage

### Known limitations

N/A