Add a function to invoke host function 'end-to-end' in recording mode. (#1332)
### What

In this mode we always use the recording mode for storage and optionally
used recording auth.

Also added some basic test coverage for all the host function kinds, as
well as some tests for simulation with more complex function calls
(involving multi-party auth).

The new function makes the best effort at emulating work performed in
the enforcing mode and while it's still imperfect, the difference is <1%
for the tests.

### Why

This function will be used to improve the accuracy soroban-simulation.

### Known limitations

N/A

---------

Co-authored-by: Siddharth Suresh <siddharth@stellar.org>