Small fixes in soroban-simulation (#1361)
### What

- Use `&impl SnapshotSourceWithArchive` in `load_from_snapshot`. This is
not very pretty semantically, but there is no clean way for storage to
implement archived entry filtering without accessing the config and I
really want to encapsulate the config-related logic in `NetworkConfig`
struct. The alternative would be for `SnapshotSource` implementer to
either load the state expiration config before building the config, or
to not follow the interface contract of `SnapshotSource` to not ever
return expired entries.
- Use `&self` in `fill_config_fields_in_ledger_info`

### Why

Issue fixes

### Known limitations

N/A