Module cache followup (#1375)
This is a set of followups to #1359 to cover missing testing and wiring
around the module cache. It includes calibration of the
`VmCachedInstantiation` cost type, a shift to using a single shared
linker when using a module cache (this turns out to be critical -- the
existing code deadlocks on cross-contract calls) and some fairly subtle
wiring to patch up costs in recording mode, along with some tests to
check that it's mostly working.

Several tests remain broken with --feature=next, so I wouldn't say it's
quite ready yet. I tried, but am out of time for tonight.