Improved TTL testing UX and cleaned up storage diagnostics. (#1412)
### What

Improved TTL testing UX and cleaned up storage diagnostics.

Both improvements rely on passing host into storage methods. This also
allows us to use the ledger info and to encapsulate detailed diagnostic
errors in the storage module. We have to still keep the old storage
methods around. There are a few users of storage getters, which probably
don't have to use it, but the clean up will have some time and can only
be performed in v22 due to semver.

There is a subtle change in observations because we now only allocate a
new object in diagnostic mode (charged to shadow budget). However, this
shouldn't cause divergence (because this version of env won't run in
p20) and is highly unlikely to affect replay of p20 with the new env. If
necessary, we can still emulate the p20 behavior.

The TTL testing support includes:

- Provide testutil getters for the `live_until_ledger` of all the ledger
entries
- Treat expired temp entries as non-existent
- Emit special test-only errors for accessing expired persistent entries

### Why

Improving testing UX

### Known limitations

N/A