Remove rust-analyzer-compat ci check (#1427)
### What

Remove rust-analyzer-compat ci check.

### Why

We added this when we seemed to be frequently breaking rust-analyzer
which slowed down dev envs for devs. But it's been more than 1 year
since I've seen anything break rust-analyzer. RA continues to get
better, and I don't think we need to keep running this check. There have
been at times problems with the check, although not frequent, it's part
of our process that isn't serving the need it was created for any
longer.