Use explicit protocol version for test Wasms.
Also modified the makefile to regenerate Wasms in the directory corresponding to the necessary protocol version.

I'm not sure yet if we need to maintain several versions of every Wasm, so I'm only leaving the v20 Wasms checked-in for now (I did try to generate v21 though just for a sanity check). Going forward I think we could keep explicitly versioned Wasms around for tests that need them, but otherwise have just a single version.