Just upgrade to protocol 22 (#1444)
This does the most minimal set of changes to the host -- switching it
from protocol 21 to 22 -- and produces a new set of observations with
that change. Note that it's done in two steps, first by making an
_exact_ copy of the protocol 21 observations and then by updating them
after the change with a re-observation. You can observe [this
change](36bf7b742d1ab40ef2d90ef103e08e90217d3fc6) to see that the only
things that changed are what you'd expect from the minimal change: some
object hashes (and therefore footprints and such) of contracts.

The idea here is to merge this change, which is obviously semantically
benign, _before_ we land anything that makes a substantive change to the
semantics of protocol 22, so that we have a better chance of isolating
"what changed" in subsequent re-observations of the changed protocol 22.