Pure wasmi 036 (#1442)
This is a _much_ simpler alternative to #1434 -- though it cribs some
code from there -- focused on the cruder strategy of "just ship two
sorobans in core", with soroban 22 not being able to replay soroban 21
contracts precisely, but soroban 21 being linked-in to core to handle
such replay.

This makes the task here dramatically simpler -- just adapt to API
changes in wasmi and recalibrate -- with a certain amount more work to
do on the core side, but not a ton.