Rounding improvement for budget model with `iterations > 1` (#1450)
### What

This is a simple improvement to the rounding of budget model evaluation,
when the model is linear and iteration > 1.

Current: 1. evaluates the linear equation, 2. scale the result back,
then 3. multiply by iteration.
This change: 1. evaluates the linear equation, 2. multiply by iteration
3. scale the result back.

The latter results in less rounding and better precision when the
iteration count is large.
Also added a test to illustrated expected behavior.

### Why

[TODO: Why this change is being made. Include any context required to
understand the why.]

### Known limitations

[TODO or N/A]

---------

Co-authored-by: Siddharth Suresh <siddharth@stellar.org>