Call constructors for the test contracts. (#1455)
### What

Call constructors for the test contracts. Also added some utilities to
support setting up contracts with constructors in the SDK.

### Why

Supporting constructors in the test framework.

### Known limitations

I had to reduce the number of iterations in
`check_caller_and_callee_seed_always_different` test as the observations
library chokes due to too many observations. I'm not sure what's the
reason, but the bug and fix are unrelated to this PR.