Bump tracy-client to 0.17 to expose delayed-init feature (#1462)
This bumps the tracy-client crate to 0.17, so that we can access the
`delayed-init` feature (it's not present in 0.16.4 but necessary to use
with core). Previously we were patching this dep entirely with a local
vendored build of tracy in core, so this potentially moves us closer to
upstream (if it work).